improved business logic of project

+++ src/components/MainContainer.jsx
@@ -4,32 +4,34 @@ import React, { useState } from 'react';
 export default function MainContainer () {
   const [currentUser, setCurrentUser] = useState('725543eb-8fd4-4e43-b5ac-2374c16900ef');
   const [currentTask, setCurrentTask] = useState('');
+  const [currentTaskComplete, setCurrentTaskComplete] = useState(true);
 
   const handleInput = () => {
     const reqOptions = {
       method: 'POST',
       headers: { 'Content-Type': 'application/json' }
     };
-    fetch(`api/${currentUser}/${currentTask.replace(' ', '_')}`, reqOptions);
+    fetch(`api/${currentUser}/${currentTask.replace(' ', '_')}`, reqOptions)
+      .then(setCurrentTaskComplete(false));
   };
 
   const handleComplete = () => {
-    if (currentTask !== '') {
+    if (currentTask !== '' && currentTaskComplete === false) {
       const reqOptions = {
         method: 'PATCH',
         headers: { 'Content-Type': 'application/json' }
       };
-      fetch(`api/${currentUser}`, reqOptions);
+      fetch(`api/${currentUser}`, reqOptions).then(setCurrentTaskComplete(true));
     };
   };
 
   const handleDelete = () => {
-    if (currentTask !== '') {
+    if (currentTask !== '' && currentTaskComplete === false) {
       const reqOptions = {
         method: 'DELETE',
         headers: { 'Content-Type': 'application/json' }
       };
-      fetch(`api/${currentUser}`, reqOptions);
+      fetch(`api/${currentUser}`, reqOptions).then(setCurrentTaskComplete(true));
     }
   };
